+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Transaction call object to accept both `input` and `data` field simultaneously if they are set to equal values [#6702](https://github.com/hyperledger/besu/pull/6702)
 
 ### Bug fixes
+- Make block transaction selection max time aware of PoA transictions [#6676](https://github.com/hyperledger/besu/pull/6676)
 
 ### Download Links